requirements.txt
Flask==1.1.1
gunicorn==20.0.0
Jinja2==2.10.3
lxml==4.4.1
request==2019.4.13
requests==2.22.0
requests-html==0.10.0
Werkzeug==0.16.0
flask-bootstrap4
flask_sslify==0.1.5
flask-fontawesome==0.1.2
